requirements.txt
requests==2.28.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy-Utils==0.38.3
Flask-Migrate==2.6.0
psycopg2-binary==2.9.3
Flask==2.2.2
python-decouple==3.6